Added old snapshot and working POD
POD is working now since I added an old snapshot matrix with working Fourier basis. The current version of Fourier is not working correctly which is causing bugs. Something in the reduction of for loops must have gone wrong. So I am going to try to make a state space implementation.